+++ examples/08-dynamic-blocks/main.tf
@@ -1,0 +1,15 @@
+locals {
+  generated_rules = [
+    for rule in var.network_rules : {
+      name     = rule.name
+      priority = rule.priority
+      action   = rule.action
+    }
+  ]
+}
+
+resource "terraform_data" "network_configuration" {
+  input = {
+    rules = local.generated_rules
+  }
+}